addressed preston comments
